+++ core/src/br/com/orlandoburli/framework/core/be/validation/utils/CPFUtils.java
@@ -22,7 +22,7 @@ public final class CPFUtils {
 				// converte o i-esimo caractere do CPF em um numero:
 				// por exemplo, transforma o caractere '0' no inteiro 0
 				// (48 eh a posicao de '0' na tabela ASCII)
-				num = (int) (CPF.charAt(i) - 48);
+				num = CPF.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso - 1;
 			}
@@ -38,7 +38,7 @@ public final class CPFUtils {
 			sm = 0;
 			peso = 11;
 			for (i = 0; i < 10; i++) {
-				num = (int) (CPF.charAt(i) - 48);
+				num = CPF.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso - 1;
 			}
